+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes in gin-zerologger will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.3] - 2025-02-28
+
+- Fixed an issue when global logging level conflicts with level-specific logging, log messages were produced unexpectedly:
+  - for example, when global logging level is set to `info` while a 200-level logging is set to `debug`, the logger will suppress any 200-level log messages (thus honoring the `info` global logging level)
+- Enhancement added to the example to support an optionally provided global logging level via an environment variable `LOGGING_LEVEL`
+
 ## [1.3.2] - 2023-12-06
 
 - Fixed issue where log level override for 400s and 500s was not being applied correctly